Properly remove spawner when used or dug
Using minetest.dig_node causes various issues. An example is that when
area protection is in effect, it fails (even for your own area, because
minetest.dig_node doesn't allow a player to be specfied, it uses a null
one). So in the event of failure, a) after spawning, the spawner stays
there, and b) when you dig it, you get a copy in your inventory (because
the return value from dig_node was ignored) but it stays there so you
can dig infinite amounts of it.

Just using remove_node resolves all this.